requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
cfgv==3.4.0
distlib==0.3.9
filelock==3.16.1
frozenlist==1.5.0
identify==2.6.2
idna==3.10
iso8601==2.1.0
multidict==6.1.0
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
propcache==0.2.0
python-dotenv==1.0.1
python-logging-mixin==0.1.3
PyYAML==6.0.2
SQLAlchemy==2.0.36
twitchio==2.10.0
typing_extensions==4.12.2
virtualenv==20.27.1
yarl==1.18.0
